Safari-609.1.11.1/LayoutTests/http/tests/security/xss-DENIED-xsl-document-expected.txt

CONSOLE MESSAGE: Unsafe attempt to load URL http://localhost:8000/security/resources/target.xml from origin http://127.0.0.1:8000. Domains, protocols and ports must match.

This test includes content via a cross-origin document() command. It passes if the load fails and thus there is no text below this line.
